autoapply/requirements.txt
pyyaml>=6.0
jinja2>=3.1
requests>=2.31
beautifulsoup4>=4.12
dnspython>=2.6
google-api-python-client>=2.100
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.1
colorama>=0.4
